Extract Or::* classes into separate files
